change mistakes

+++ src/Itmo/lessons1/Task2.java
@@ -2,7 +2,7 @@ package Itmo.lessons1;
 
 public class Task2 {
     public double expressionResult1() {
-        double result1 = ( 46 + 10 ) * ( 10 / 3 );
+        double result1 = ( 46 + 10 ) * ((double) 10 / 3);
         return result1;
 
     }
