+++ src/Itmo/lessons1/Task3.java
@@ -3,7 +3,7 @@ package Itmo.lessons1;
 public class Task3 {
     public void count(){
         int number = 10500;
-        int result = (number/10)/10;
+        double result = ((double) (number/10)/10);
         System.out.println(result);
     }
 }
